requirements.txt
python-telegram-bot
docker